[expo-router] migrate native-stack to standard-navigator
