+++ packages/expo-router/CHANGELOG.md
@@ -38,6 +38,7 @@
 - Move `@testing-library/jest-dom` and `@testing-library/user-event` from `dependencies` to `devDependencies` ([#47820](https://github.com/expo/expo/pull/47820) by [@krystofwoldrich](https://github.com/krystofwoldrich))
 - Rewrite native tabs using standard-navigation ([#46457](https://github.com/expo/expo/pull/46457) by [@Ubax](https://github.com/Ubax))
 - [Internal] Split `useLoaderData()` into a document cache and a per-mount Suspense store ([#47365](https://github.com/expo/expo/pull/47365) by [@hassankhan](https://github.com/hassankhan))
+- Migrate the native `Stack` navigator to the `standard-navigation` integration. ([@Ubax](https://github.com/Ubax))
 
 ## 56.2.7 — 2026-05-26
