+++ packages/expo-router/src/react-navigation/native-stack/index.tsx
@@ -1,9 +1,4 @@
 /**
- * Navigators
- */
-export { createNativeStackNavigator } from './navigators/createNativeStackNavigator';
-
-/**
  * Views
  */
 export { NativeStackView } from './views/NativeStackView';
