--- packages/expo-router/src/react-navigation/native-stack/navigators/createNativeStackNavigator.tsx
@@ -1,124 +1,0 @@
-'use client';
-import * as React from 'react';
-import { use } from 'react';
-
-import {
-  createNavigatorFactory,
-  type EventArg,
-  NavigationMetaContext,
-  type NavigatorTypeBagBase,
-  type ParamListBase,
-  type StackActionHelpers,
-  StackActions,
-  type StackNavigationState,
-  StackRouter,
-  type StackRouterOptions,
-  type StaticConfig,
-  type TypedNavigator,
-  useNavigationBuilder,
-} from '../../native';
-import type {
-  NativeStackNavigationEventMap,
-  NativeStackNavigationOptions,
-  NativeStackNavigationProp,
-  NativeStackNavigatorProps,
-} from '../types';
-import { usePopAction } from '../utils/usePopAction';
-import { useProjectedStack } from '../utils/useProjectedStack';
-import { NativeStackView } from '../views/NativeStackView';
-
-function NativeStackNavigator({
-  id,
-  initialRouteName,
-  UNSTABLE_routeNamesChangeBehavior,
-  children,
-  layout,
-  screenListeners,
-  screenOptions,
-  screenLayout,
-  UNSTABLE_router,
-  ...rest
-}: NativeStackNavigatorProps) {
-  const { state, describe, descriptors, navigation, NavigationContent } = useNavigationBuilder<
-    StackNavigationState<ParamListBase>,
-    StackRouterOptions,
-    StackActionHelpers<ParamListBase>,
-    NativeStackNavigationOptions,
-    NativeStackNavigationEventMap
-  >(StackRouter, {
-    id,
-    initialRouteName,
-    UNSTABLE_routeNamesChangeBehavior,
-    children,
-    layout,
-    screenListeners,
-    screenOptions,
-    screenLayout,
-    UNSTABLE_router,
-  });
-
-  const meta = use(NavigationMetaContext);
-
-  React.useEffect(() => {
-    if (meta && 'type' in meta && meta.type === 'native-tabs') {
-      // If we're inside native tabs, we don't need to handle popToTop
-      // It's handled natively by native tabs
-      return;
-    }
-
-    // @ts-expect-error: there may not be a tab navigator in parent
-    return navigation?.addListener?.('tabPress', (e: any) => {
-      const isFocused = navigation.isFocused();
-
-      // Run the operation in the next frame so we're sure all listeners have been run
-      // This is necessary to know if preventDefault() has been called
-      requestAnimationFrame(() => {
-        if (state.index > 0 && isFocused && !(e as EventArg<'tabPress', true>).defaultPrevented) {
-          // When user taps on already focused tab and we're inside the tab,
-          // reset the stack to replicate native behaviour
-          navigation.dispatch({
-            ...StackActions.popToTop(),
-            target: state.key,
-          });
-        }
-      });
-    });
-  }, [meta, navigation, state.index, state.key]);
-
-  // Project preloaded routes as regular routes after `index`, with descriptors covering them.
-  // The view then treats any route positioned after the focused one as preloaded.
-  const { projectedState, projectedDescriptors } = useProjectedStack(state, descriptors, describe);
-
-  const pop = usePopAction(navigation, state.key);
-
-  return (
-    <NavigationContent>
-      <NativeStackView
-        {...rest}
-        state={projectedState}
-        descriptors={projectedDescriptors}
-        emit={navigation.emit}
-        pop={pop}
-      />
-    </NavigationContent>
-  );
-}
-
-export function createNativeStackNavigator<
-  const ParamList extends ParamListBase,
-  const NavigatorID extends string | undefined = string | undefined,
-  const TypeBag extends NavigatorTypeBagBase = {
-    ParamList: ParamList;
-    NavigatorID: NavigatorID;
-    State: StackNavigationState<ParamList>;
-    ScreenOptions: NativeStackNavigationOptions;
-    EventMap: NativeStackNavigationEventMap;
-    NavigationList: {
-      [RouteName in keyof ParamList]: NativeStackNavigationProp<ParamList, RouteName, NavigatorID>;
-    };
-    Navigator: typeof NativeStackNavigator;
-  },
-  const Config extends StaticConfig<TypeBag> = StaticConfig<TypeBag>,
->(config?: Config): TypedNavigator<TypeBag, Config> {
-  return createNavigatorFactory(NativeStackNavigator)(config);
-}
